+++ packages/editor-common/web/src/index.ts
@@ -137,7 +137,6 @@ import DraftOffsetKey from '@wix/draft-js/lib/DraftOffsetKey';
 export { DraftOffsetKey };
 
 export { elementOverflowWithEditor } from './Utils/overflowUtils';
-export { createEditorStyles } from './Utils/createEditorStyles';
 
 export const getEmptyDraftContent: () => DraftContent = () => ({
   entityMap: {},