+++ packages/tiptap-editor/web/src/components/RicosTiptapEditor/core-configs.ts
@@ -11,6 +11,12 @@ import {
   createParagraph,
   createItalic,
   createUnderline,
+  createText,
+  createDoc,
+  createListItem,
+  createBlockquote,
+  createOrderedList,
+  createTrailingNode,
 } from 'wix-tiptap-extensions';
 
 export const coreConfigs: RicosExtension[] = [
@@ -25,4 +31,10 @@ export const coreConfigs: RicosExtension[] = [
   createParagraph(),
   createItalic(),
   createUnderline(),
+  createText(),
+  createDoc(),
+  createListItem(),
+  createBlockquote(),
+  createOrderedList(),
+  createTrailingNode(),
 ];